split oversized outage proof helper (#1963)
