+++ app/src/androidTest/java/com/pocketshell/app/proof/OutboundExactlyOnceAcrossFlapE2eTest.kt
@@ -1208,25 +1208,6 @@ class OutboundExactlyOnceAcrossFlapE2eTest {
     } }
 
     /**
-     * The text the fake agent SUBMITTED, whitespace-stripped — everything between
-     * its `FAKE-AGENT SUBMITTED: ` marker and the `> ` input box that follows it.
-     * The input box wraps across rows as it grows, so the capture is compared
-     * whitespace-insensitively; the payload carries no whitespace of its own, so
-     * this is exact for the property under test.
-     */
-    private fun submittedTextStripped(capture: String): String {
-        val stripped = capture.filterNot { it.isWhitespace() }
-        val start = stripped.indexOf(FAKE_AGENT_SUBMITTED_STRIPPED)
-        assertTrue(
-            "the fixture must have submitted a prompt; capture:\n$capture",
-            start >= 0,
-        )
-        val body = stripped.substring(start + FAKE_AGENT_SUBMITTED_STRIPPED.length)
-        val inputBox = body.indexOf('>')
-        return if (inputBox >= 0) body.substring(0, inputBox) else body
-    }
-
-    /**
      * Keystroke lane: input typed through the REAL TerminalView session (pane
      * input queue + pump) whose first send loses its result must NOT be doubled
      * by the retry — the probe sees it landed and suppresses attempt 2.
@@ -1741,20 +1722,6 @@ class OutboundExactlyOnceAcrossFlapE2eTest {
             inputLine,
         )
     }
-
-    private fun countOccurrences(haystack: String, needle: String): Int {
-        if (needle.isEmpty()) return 0
-        var count = 0
-        var index = haystack.indexOf(needle)
-        while (index >= 0) {
-            count += 1
-            index = haystack.indexOf(needle, index + needle.length)
-        }
-        return count
-    }
-
-    private fun countCollapsedPasteChips(capture: String): Int =
-        Regex("""\[Pasted text #\d+ \+\d+ lines?]""").findAll(capture).count()
 
     // ---------------------------------------------------------------- attach helpers
 
@@ -2548,7 +2515,6 @@ class OutboundExactlyOnceAcrossFlapE2eTest {
         const val FAKE_AGENT_LEDGER_PATH: String = "/tmp/pocketshell-fake-agent-submits.log"
         const val SEEDED_JSONL: String = "issue1526-live-claude.jsonl"
         const val FAKE_AGENT_READY: String = "FAKE-AGENT-READY"
-        const val FAKE_AGENT_SUBMITTED_STRIPPED: String = "FAKE-AGENTSUBMITTED:"
         const val CONVERSATION_SEGMENT_TAG: String = TMUX_CONSOLIDATED_TAB_PILL_TAG_PREFIX + "1"
 
         val DETECTION_TIMEOUT_MS: Long =
